Refactored new code into new file
